requirements.txt
pandas>=2.0
numpy>=1.24
openpyxl>=3.1
gspread>=6.0
google-auth>=2.22
google-auth-oauthlib>=1.1
schedule>=1.2
faker>=20.0
